Fix the root cause of killing everything
From kill's manpage:
 -1     All processes with a PID larger than 1 are signaled

Authored-by: Chen Mulong <muchen@pivotal.io>

+++ src/backend/cdb/endpoint/cdbendpointretrieve.c
@@ -563,9 +563,12 @@ retrieve_cancel_action(const char *endpointName, char *msg)
 	{
 		endpointDesc->receiver_pid = InvalidPid;
 		endpointDesc->attach_status = Status_Released;
-		elog(DEBUG3, "CDB_ENDPOINT: signal sender to abort");
-		SetBackendCancelMessage(endpointDesc->sender_pid, msg);
-		kill(endpointDesc->sender_pid, SIGINT);
+		if (endpointDesc->sender_pid != InvalidPid)
+		{
+			elog(DEBUG3, "CDB_ENDPOINT: signal sender to abort");
+			SetBackendCancelMessage(endpointDesc->sender_pid, msg);
+			kill(endpointDesc->sender_pid, SIGINT);
+		}
 	}
 
 	LWLockRelease(ParallelCursorEndpointLock);
